fix: convert timestamp to bigint for improved precision

+++ src/pages/api/status.ts
@@ -26,7 +26,7 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse<
     return res.status(200).json({
       status: shouldReturnTrue,
       message: shouldReturnTrue ? "Operation successful" : "Operation failed",
-      timestamp: Math.floor(Date.now() / 1000), // Unix timestamp as uint32
+      timestamp: BigInt(Math.floor(Date.now() / 1000)), // Unix timestamp as uint32
     });
   } catch (error) {
     const errorMessage = error instanceof Error ? error.message : "Unknown error";
